ModuleRunner enhancements
 - Move the launcher property `modules` to `ModuleLauncherProperties`
   - This will make the `ModuleRunner` to be `Component` bean only; Also, it only needs to enable the configuration properties of `ModuleLauncherProperties`

Separate module resolver/launcher properties

Polishing

Add validator dependency to module launcher
